+++ Arbiter.App/Models/WorldEntity.cs
@@ -1,3 +1,0 @@
-﻿namespace Arbiter.App.Models;
-
-public record WorldEntity(WorldEntityType Type, uint Id, int X, int Y, int Sprite, string? Name);