+++ Arbiter.App/ViewModels/Proxy/ProxyViewModel.ModMenuFilters.cs
@@ -1,0 +1,6 @@
+﻿namespace Arbiter.App.ViewModels.Proxy;
+
+public partial class ProxyViewModel
+{
+    
+}